new version of shift array
